+++ src/main/resources/i18n/messages.properties
@@ -1788,16 +1788,15 @@ announcement.user.readon=Read by you
 announcement.markRead=Mark as Read
 announcement.markUnread=Mark as Unread
 
-AnnouncementTable.title=Title
+AnnouncementDetails.title=Announcement Details
 
-AnnouncementModal.create.title=Add Announcement
-AnnouncementModal.edit.title=Edit Announcement
-AnnouncementModal.create.okBtn=Create Announcement
-AnnouncementModal.edit.okBtn=Update Announcement
-AnnouncementModal.form.title=Title
-AnnouncementModal.form.message=Message
-AnnouncementModal.form.priority=High Priority?
-AnnouncementModal.form.error.title=Please enter a title
+AnnouncementForm.create.button=Create Announcement
+AnnouncementForm.edit.button=Update Announcement
+AnnouncementForm.delete.button=Delete Announcement
+AnnouncementForm.form.title=Title
+AnnouncementForm.form.message=Message
+AnnouncementForm.form.priority=High Priority?
+AnnouncementForm.form.error.title=Please enter a title
 
 CreateNewAnnouncement.title=Add Announcement
