+++ src/main/webapp/resources/js/pages/admin/components/announcements/AnnouncementUserTable.jsx
@@ -3,6 +3,11 @@ import { PagedTable } from "../../../../components/ant.design/PagedTable";
 import { formatDate } from "../../../../utilities/date-utilities";
 import { setBaseUrl } from "../../../../utilities/url-utilities";
 
+/**
+ * Render React component to show the user views of an announcement.
+ * @returns {*}
+ * @constructor
+ */
 export default function AnnouncementUserTable() {
   const columns = [
     {
