+++ src/main/webapp/resources/js/pages/admin/components/announcements/DeleteAnnouncement.jsx
@@ -1,12 +1,15 @@
 import React from "react";
 import { Button, Popconfirm } from "antd";
 import { red6 } from "../../../../styles/colors";
-import { IconQuestionCircle, IconTrash } from "../../../../components/icons/Icons";
+import {
+  IconQuestionCircle,
+  IconTrash,
+} from "../../../../components/icons/Icons";
 
 /**
  * Component to render a Delete announcement button.
- * @param id
- * @param deleteAnnouncement
+ * @param {integer} id - the id of the announcement that is to be deleted.
+ * @param {function} deleteAnnouncement - the function that deletes an announcement.
  * @returns {*}
  * @constructor
  */
